MAKE app-info.php helper file that's included in laravel.php
REFACTOR environment & version helpers to app-info.php.
